Convert camps to nested_set
Technically, all events are now in a nested set, so we have a lot
of root nodes. Since we control which events are nested, this may
not be a big problem.

If you stumble upon this commit while debugging something event-
related: I hold no emotional value to this implementation. I would
really like to confine the nested-set to Event::Camp, but did
not succeed while trying. May  be with you in you quest of
bettering the code.