Fix creating event roles for people who already have one
The attributes of the event participations and event roles have been
assigned in a nonstandard way since 2012. Unfortunately, this means we
cannot just extend the available fields in the normal way (by adding to
the permitted_attrs). Instead, we have to extend the build_entry method
so that it sets the fields.
Extending the permitted_attrs works when only adding completely new
people to the event. But when adding another role for a person who
already had a participation in the event, the assign_attributes call
would completely overwrite the role's participation which has previously
been carefully crafted in the build_entry method.